requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
requests==2.25.0
six==1.15.0
spotipy==2.16.1
urllib3==1.26.2
python-decouple==3.3
matplotlib==3.3.2
more-itertools==8.6.0
pandas==1.1.5
